server/requirements.txt
# Core dependencies
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
supabase==2.10.0
websockets>=12,<14
google-auth>=2.0.0,<3

# Optional for development
# pytest==8.3.3
# black==24.10.0
# mypy==1.11.2
